Add parsing check: Direction vectors need to be normalized. To check if a vector is normalized, this formula helps:     cos(alpha)^2 + cos(beta)^2 + cos(gamma)^2 = 1
But since, we are dealing with a vector with coordinates from -1 to 1,
we can replace cosines by the coordinates of the vector:
    v.x^2 + v.y^2 + v.z^2 = 1

This formula sometimes needs extremely precise coordinates to make the formula true.
This is why I decided to modify the formula to be more permissive:
    abs(v.x^2 + v.y^2 + v.z^2 - 1) <= 0.01

This ensures that the parsed vector is almost normalized for a better user experience.